# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
security:
  - clerkSession: []
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Return tasks scheduled for this local calendar date.
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [today, inbox, all]
            default: today
        - name: timezone
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
          description: IANA timezone used to interpret the local calendar date.
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /settings/notifications:
    get:
      operationId: getNotificationSettings
      tags: [notifications]
      summary: Read notification preferences
      description: Auto-creates defaults on first read.
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
    patch:
      operationId: updateNotificationSettings
      tags: [notifications]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSettingsUpdate"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
        "400":
          description: Invalid preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /settings/focus:
    get:
      operationId: getFocusSettings
      tags: [focus]
      summary: Read focus preferences
      description: Auto-creates defaults on first read.
      responses:
        "200":
          description: Focus preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSettings"
    patch:
      operationId: updateFocusSettings
      tags: [focus]
      summary: Update focus preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FocusSettingsUpdate"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSettings"
        "400":
          description: Invalid preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /tasks/summary:
    get:
      operationId: getTaskSummary
      tags: [tasks]
      summary: Get daily task summary
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: timezone
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
          description: IANA timezone used to interpret the local calendar date.
      responses:
        "200":
          description: Daily task summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSummary"
...
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Clerk session token for the signed-in user. Browsers send it via the
        Clerk session cookie through @clerk/express; non-browser clients send
        `Authorization: Bearer <session JWT>`. Every route except GET
        /healthz returns 401 without it.
  schemas:
...
    Task:
      type: object
      required: [id, title, notes, dueAt, durationMin, priority, status, projectId, tags, parentId, rescheduleCount, needsAttention, automation, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        notes:
          type: ["string", "null"]
        dueAt:
          type: ["string", "null"]
          format: date-time
        durationMin:
          type: integer
          minimum: 5
          maximum: 1440
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [inbox, open, completed]
        projectId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        parentId:
          type: ["integer", "null"]
        rescheduleCount:
          type: integer
        needsAttention:
          type: boolean
        automation:
          type: ["string", "null"]
          enum: [off, ask, auto]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 240
        notes:
          type: ["string", "null"]
          maxLength: 4000
        dueAt:
          type: ["string", "null"]
          format: date-time
        durationMin:
          type: integer
          minimum: 5
          maximum: 1440
          default: 30
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        status:
          type: string
          enum: [inbox, open]
          default: open
        projectId:
          type: ["integer", "null"]
          description: >-
            Owning project id. Must belong to the caller (else 404).
            Null or omitted leaves the task unfiled.
        tagIds:
          type: array
          items:
            type: integer
          maxItems: 20
          description: >-
            Tag ids to attach. Every id must belong to the caller (else 404).
        parentId:
          type: ["integer", "null"]
          description: >-
            Parent task id for subtasks. Must belong to the caller (else 404).
            Null or omitted creates a top-level task.
        automation:
          type: ["string", "null"]
          enum: [off, ask, auto]
          description: >-
            Per-task automation dial override. Null or omitted inherits the
            user's rescheduling default mode.
        dueText:
          type: ["string", "null"]
          maxLength: 120
          description: >-
            Natural-language due date ("tomorrow 5pm", "fri", "sep 20 9am").
            Resolved server-side in `timezone` and stored as `dueAt`.
            Explicit `dueAt` and `dueText` are mutually exclusive;
            unparseable text is rejected with 400, never silently dropped.
        timezone:
          type: string
          description: >-
            IANA time-zone identifier used to interpret `dueText`
            (e.g. "Asia/Kolkata"). Invalid or absent values fall back to UTC.
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 240
        notes:
          type: ["string", "null"]
          maxLength: 4000
        dueAt:
          type: ["string", "null"]
          format: date-time
        durationMin:
          type: integer
          minimum: 5
          maximum: 1440
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [inbox, open, completed]
        projectId:
          type: ["integer", "null"]
          description: >-
            Owning project id (must belong to the caller, else 404).
            Null clears the filing; omitted leaves it unchanged.
        tagIds:
          type: array
          items:
            type: integer
          maxItems: 20
          description: >-
            Tag ids to attach. A present array REPLACES the full set
            (            empty array clears all); omitted or null keeps it.
            Every id must belong to the caller (else 404).
        parentId:
          type: ["integer", "null"]
          description: >-
            Parent task id. Must belong to the caller (else 404); cyclic
            assignments are rejected with 400. Null detaches to top level;
            omitted leaves it unchanged.
        automation:
          type: ["string", "null"]
          enum: [off, ask, auto]
          description: >-
            Per-task dial override. Null clears back to the user's default
            mode; omitted leaves it unchanged.
        dueText:
          type: ["string", "null"]
          maxLength: 120
          description: >-
            Natural-language due date, resolved server-side like on create.
            `null` leaves `dueAt` unchanged (send explicit `dueAt: null` to
            clear it). `dueAt` and `dueText` are mutually exclusive.
        timezone:
          type: string
          description: >-
            IANA time-zone identifier used to interpret `dueText`.
            Invalid or absent values fall back to UTC.
...
    Tag:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    NotificationSettings:
      type: object
      required: [telegramChatId, quietStart, quietEnd, timezone, remindersEnabled, createdAt, updatedAt]
      properties:
        telegramChatId:
          type: ["string", "null"]
        quietStart:
          type: integer
          minimum: 0
          maximum: 23
        quietEnd:
          type: integer
          minimum: 0
          maximum: 23
        timezone:
          type: string
        remindersEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NotificationSettingsUpdate:
      type: object
      properties:
        telegramChatId:
          type: ["string", "null"]
          pattern: '^-?[0-9]{1,19}$'
          description: Numeric Telegram chat id as text. Null unlinks.
        quietStart:
          type: integer
          minimum: 0
          maximum: 23
        quietEnd:
          type: integer
          minimum: 0
          maximum: 23
        timezone:
          type: string
          description: IANA identifier; invalid values are rejected with 400.
        remindersEnabled:
          type: boolean
    FocusSettings:
      type: object
      required: [dailyTarget, createdAt, updatedAt]
      properties:
        dailyTarget:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FocusSettingsUpdate:
      type: object
      properties:
        dailyTarget:
          type: integer
          minimum: 1
          maximum: 20
...
    TaskSummary:
      type: object
      required: [total, completed, open, focusMinutes]
      properties:
        total:
          type: integer
        completed:
          type: integer
        open:
          type: integer
        focusMinutes:
          type: integer
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string