# lib/api-zod/src/generated/types/notificationSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NotificationSettings {
  /** @nullable */
  telegramChatId: string | null;
  /**
     * @minimum 0
     * @maximum 23
     */
  quietStart: number;
  /**
     * @minimum 0
     * @maximum 23
     */
  quietEnd: number;
  timezone: string;
  remindersEnabled: boolean;
  createdAt: Date;
  updatedAt: Date;
}